Use separated resolver layer for AR
