+++ lib/clowne/adapters/active_record.rb
@@ -15,4 +15,4 @@ ActiveSupport.on_load(:active_record) do
 end
 
 require 'clowne/adapters/active_record/associations'
-require 'clowne/adapters/active_record/association'
+require 'clowne/adapters/active_record/resolvers/association'
