+++ lib/clowne/adapters/resolvers/nullify.rb
@@ -2,7 +2,7 @@
 
 module Clowne
   module Adapters
-    class Base
+    class Resolvers
       module Nullify # :nodoc: all
         def self.call(_source, record, declaration, **_options)
           declaration.attributes.each do |attr|
@@ -16,4 +16,4 @@ module Clowne
   end
 end
 
-Clowne::Adapters::Base.register_resolver(:nullify, Clowne::Adapters::Base::Nullify)
+Clowne::Adapters::Base.register_resolver(:nullify, Clowne::Adapters::Resolvers::Nullify)
